refactor(runtimed): split peer eviction handling
